+++ ImageProcessing/Source/ImageProcessingInterface.h
@@ -48,7 +48,7 @@ struct frame;
 class ImageProcessingInterface
 {
 public:
-	virtual  void ThresholdMethod(const frame& Input, frame& Output, const float ThresholdValue) = 0;
+	virtual  void ThresholdMethod(const frame& Input, frame& Output, const uint8_t ThresholdValue) = 0;
 	virtual void MotionDetection(const frame& Input, frame& Output) = 0;
 	virtual void LowpassFilter(const frame& Input, frame& Output, const float LowPassRC) = 0;
 	virtual void ApplyKernel(const frame& Input, frame& Output, const std::vector<float> Kernel) = 0;
@@ -56,6 +56,6 @@ public:
 	virtual void MorphOperation(const frame& Input, frame& Output, IImgaeProcessor::MORPHOP operation, int MorphCount) = 0;
 	virtual void MedianFilter(const frame& Input, frame& Output) = 0;
 	virtual void Adaptive(const frame& Input, frame& Output, const float AdaptiveBias) = 0;
-	virtual void Otsu(const frame& Input, frame& Output, float& ThresholdValue) = 0;
+	virtual void Otsu(const frame& Input, frame& Output, uint8_t& ThresholdValue) = 0;
 	virtual void Rotation(const frame& Input, frame& Output, IImgaeProcessor::INTERPOLATION Interpol, const float Angle) = 0;
 };
